% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_facts.R
\name{get_facts_version}
\alias{get_facts_version}
\title{Get FACTS version matching your FACTS file}
\usage{
get_facts_version(facts_file)
}
\arguments{
\item{facts_file}{Character, name of a FACTS file.
Usually has a \verb{*.facts} file extension.}
}
\value{
A version string.
}
\description{
Get the version of FACTS compatible with
your \verb{*.facts} file.
}
\examples{
facts_file <- get_facts_file_example("contin.facts")
facts_file
get_facts_version(facts_file)
}
\seealso{
\code{\link[=get_facts_versions]{get_facts_versions()}}
}
