% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{get_param_files}
\alias{get_param_files}
\title{List the paths to the param files}
\usage{
get_param_files(param_files)
}
\arguments{
\item{param_files}{Character vector of directories containing param files.}
}
\value{
Character vector of paths to param files.
}
\description{
List the paths to the all the param files
in a directory or directories.
}
\examples{
# Can only run if system dependencies are configured:
if (file.exists(Sys.getenv("RFACTS_PATHS"))) {
facts_file <- get_facts_file_example("contin.facts")
dir <- run_flfll(facts_file, verbose = FALSE)
get_param_files(dir)
}
}
